Allow more compact node notation
Instead of

    foo:
      kind: requirement
      title: Foo
      tags:
        - t/bar
        - t/baz

it is now possible to write

    foo:
      text: |
        # requirement: Foo

        +t/bar +t/baz

Of course short forms for "requirement", "component" etc. are
still possible:

    foo:
      text: |
        # r: Foo

        +t/bar +t/baz

The resulting node will be rendered with the title "Foo" and the
line will be stripped from the markdown text. Also, the fields
`text` and `description` are now synonymous. Combinations of the
old and new syntax are also supported:

    foo:
      description: |
        # r: Foo

        +t/bar
      tags:
        - t/baz